+++ packages/profile-sync-controller/src/controllers/authentication/AuthenticationController.ts
@@ -14,6 +14,7 @@ import type { HandleSnapRequest } from '@metamask/snaps-controllers';
 
 import {
   createSnapPublicKeyRequest,
+  createSnapAllPublicKeysRequest,
   createSnapSignMessageRequest,
 } from './auth-snap-requests';
 import type { LoginResponse, SRPInterface, UserProfile } from '../../sdk';
@@ -30,7 +31,7 @@ const controllerName = 'AuthenticationController';
 // State
 export type AuthenticationControllerState = {
   isSignedIn: boolean;
-  sessionData?: LoginResponse;
+  srpSessionData?: Record<string, LoginResponse>;
 };
 export const defaultState: AuthenticationControllerState = {
   isSignedIn: false,
@@ -40,7 +41,7 @@ const metadata: StateMetadata<AuthenticationControllerState> = {
     persist: true,
     anonymous: true,
   },
-  sessionData: {
+  srpSessionData: {
     persist: true,
     anonymous: false,
   },
@@ -214,25 +215,46 @@ export default class AuthenticationController extends BaseController<
     );
   }
 
-  async #getLoginResponseFromState(): Promise<LoginResponse | null> {
-    if (!this.state.sessionData) {
+  async #getLoginResponseFromState(
+    entropySourceId?: string,
+  ): Promise<LoginResponse | null> {
+    if (entropySourceId) {
+      if (!this.state.srpSessionData?.[entropySourceId]) {
+        return null;
+      }
+      return this.state.srpSessionData[entropySourceId];
+    }
+
+    const primarySrpLoginResponse = Object.values(
+      this.state.srpSessionData || {},
+    )?.[0];
+
+    if (!primarySrpLoginResponse) {
       return null;
     }
 
-    return this.state.sessionData;
+    return primarySrpLoginResponse;
   }
 
-  async #setLoginResponseToState(loginResponse: LoginResponse) {
+  async #setLoginResponseToState(
+    loginResponse: LoginResponse,
+    entropySourceId?: string,
+  ) {
     const metaMetricsId = await this.#metametrics.getMetaMetricsId();
     this.update((state) => {
-      state.isSignedIn = true;
-      state.sessionData = {
-        ...loginResponse,
-        profile: {
-          ...loginResponse.profile,
-          metaMetricsId,
-        },
-      };
+      if (entropySourceId) {
+        state.isSignedIn = true;
+        if (!state.srpSessionData) {
+          state.srpSessionData = {};
+        }
+        state.srpSessionData[entropySourceId] = {
+          ...loginResponse,
+          profile: {
+            ...loginResponse.profile,
+            metaMetricsId,
+          },
+        };
+      }
     });
   }
 
@@ -242,15 +264,26 @@ export default class AuthenticationController extends BaseController<
     }
   }
 
-  public async performSignIn(): Promise<string> {
+  public async performSignIn(): Promise<string[]> {
     this.#assertIsUnlocked('performSignIn');
-    return await this.#auth.getAccessToken();
+
+    const allPublicKeys = await this.#snapGetAllPublicKeys();
+    const accessTokens = [];
+
+    // We iterate sequentially in order to be sure that the first entry
+    // is the primary SRP LoginResponse.
+    for (const [entropySourceId] of allPublicKeys) {
+      const accessToken = await this.#auth.getAccessToken(entropySourceId);
+      accessTokens.push(accessToken);
+    }
+
+    return accessTokens;
   }
 
   public performSignOut(): void {
     this.update((state) => {
       state.isSignedIn = false;
-      state.sessionData = undefined;
+      state.srpSessionData = undefined;
     });
   }
 
@@ -261,20 +294,24 @@ export default class AuthenticationController extends BaseController<
    * @returns profile for the session.
    */
 
-  public async getBearerToken(): Promise<string> {
+  public async getBearerToken(entropySourceId?: string): Promise<string> {
     this.#assertIsUnlocked('getBearerToken');
-    return await this.#auth.getAccessToken();
+    return await this.#auth.getAccessToken(entropySourceId);
   }
 
   /**
    * Will return a session profile.
    * Logs a user in if a user is not logged in.
    *
+   * @param entropySourceId - The entropy source ID used to derive the key,
+   * when multiple sources are available (Multi-SRP).
    * @returns profile for the session.
    */
-  public async getSessionProfile(): Promise<UserProfile> {
+  public async getSessionProfile(
+    entropySourceId?: string,
+  ): Promise<UserProfile> {
     this.#assertIsUnlocked('getSessionProfile');
-    return await this.#auth.getUserProfile();
+    return await this.#auth.getUserProfile(entropySourceId);
   }
 
   public isSignedIn(): boolean {
@@ -284,15 +321,33 @@ export default class AuthenticationController extends BaseController<
   /**
    * Returns the auth snap public key.
    *
+   * @param entropySourceId - The entropy source ID used to derive the key,
+   * when multiple sources are available (Multi-SRP).
    * @returns The snap public key.
    */
-  async #snapGetPublicKey(): Promise<string> {
+  async #snapGetPublicKey(entropySourceId?: string): Promise<string> {
     this.#assertIsUnlocked('#snapGetPublicKey');
 
     const result = (await this.messagingSystem.call(
       'SnapController:handleRequest',
-      createSnapPublicKeyRequest(),
+      createSnapPublicKeyRequest(entropySourceId),
     )) as string;
+
+    return result;
+  }
+
+  /**
+   * Returns a mapping of entropy source IDs to auth snap public keys.
+   *
+   * @returns A mapping of entropy source IDs to public keys.
+   */
+  async #snapGetAllPublicKeys(): Promise<[string, string][]> {
+    this.#assertIsUnlocked('#snapGetAllPublicKeys');
+
+    const result = (await this.messagingSystem.call(
+      'SnapController:handleRequest',
+      createSnapAllPublicKeysRequest(),
+    )) as [string, string][];
 
     return result;
   }
@@ -303,9 +358,14 @@ export default class AuthenticationController extends BaseController<
    * Signs a specific message using an underlying auth snap.
    *
    * @param message - A specific tagged message to sign.
+   * @param entropySourceId - The entropy source ID used to derive the key,
+   * when multiple sources are available (Multi-SRP).
    * @returns A Signature created by the snap.
    */
-  async #snapSignMessage(message: string): Promise<string> {
+  async #snapSignMessage(
+    message: string,
+    entropySourceId?: string,
+  ): Promise<string> {
     assertMessageStartsWithMetamask(message);
 
     if (this.#_snapSignMessageCache[message]) {
@@ -316,7 +376,7 @@ export default class AuthenticationController extends BaseController<
 
     const result = (await this.messagingSystem.call(
       'SnapController:handleRequest',
-      createSnapSignMessageRequest(message),
+      createSnapSignMessageRequest(message, entropySourceId),
     )) as string;
 
     this.#_snapSignMessageCache[message] = result;