+++ packages/profile-sync-controller/src/controllers/user-storage/account-syncing/utils.test.ts
@@ -6,8 +6,6 @@ import { USER_STORAGE_VERSION, USER_STORAGE_VERSION_KEY } from './constants';
 import {
   isNameDefaultAccountName,
   mapInternalAccountToUserStorageAccount,
-  isInternalAccountFromPrimarySRPHdKeyring,
-  mapInternalAccountsListToPrimarySRPHdKeyringInternalAccountsList,
 } from './utils';
 
 describe('user-storage/account-syncing/utils', () => {
@@ -75,109 +73,6 @@ describe('user-storage/account-syncing/utils', () => {
         n: internalAccountWithCustomName.metadata.name,
         nlu: internalAccountWithCustomName.metadata.nameLastUpdatedAt,
       });
-    });
-  });
-
-  describe('isInternalAccountFromPrimarySRPHdKeyring', () => {
-    const internalAccount = {
-      address: '0x123',
-      id: '1',
-      metadata: {
-        name: `${getMockRandomDefaultAccountName()} 1`,
-        nameLastUpdatedAt: 1620000000000,
-        keyring: {
-          type: KeyringTypes.hd,
-        },
-      },
-    } as InternalAccount;
-
-    const getMessenger = jest.fn();
-    const getUserStorageControllerInstance = jest.fn();
-
-    it('should return true if the internal account is from the primary SRP and is from the HD keyring', async () => {
-      getMessenger.mockReturnValue({
-        call: jest.fn().mockResolvedValue(['0x123']),
-      });
-
-      const result = await isInternalAccountFromPrimarySRPHdKeyring(
-        internalAccount,
-        { getMessenger, getUserStorageControllerInstance },
-      );
-
-      expect(result).toBe(true);
-    });
-
-    it('should return false if the internal account is not from the primary SRP', async () => {
-      getMessenger.mockReturnValue({
-        call: jest.fn().mockResolvedValue(['0x456']),
-      });
-
-      const result = await isInternalAccountFromPrimarySRPHdKeyring(
-        internalAccount,
-        { getMessenger, getUserStorageControllerInstance },
-      );
-
-      expect(result).toBe(false);
-    });
-
-    it('should return false if the internal account is not from the HD keyring', async () => {
-      getMessenger.mockReturnValue({
-        call: jest.fn().mockResolvedValue(['0x123']),
-      });
-
-      const result = await isInternalAccountFromPrimarySRPHdKeyring(
-        {
-          ...internalAccount,
-          metadata: { keyring: { type: KeyringTypes.simple } },
-        } as InternalAccount,
-        { getMessenger, getUserStorageControllerInstance },
-      );
-
-      expect(result).toBe(false);
-    });
-  });
-
-  describe('mapInternalAccountsListToPrimarySRPHdKeyringInternalAccountsList', () => {
-    const internalAccountsList = [
-      {
-        address: '0x123',
-        id: '1',
-        metadata: {
-          name: `${getMockRandomDefaultAccountName()} 1`,
-          nameLastUpdatedAt: 1620000000000,
-          keyring: {
-            type: KeyringTypes.hd,
-          },
-        },
-      } as InternalAccount,
-      {
-        address: '0x456',
-        id: '2',
-        metadata: {
-          name: `${getMockRandomDefaultAccountName()} 2`,
-          nameLastUpdatedAt: 1620000000000,
-          keyring: {
-            type: KeyringTypes.simple,
-          },
-        },
-      } as InternalAccount,
-    ];
-
-    const getMessenger = jest.fn();
-    const getUserStorageControllerInstance = jest.fn();
-
-    it('should return a list of internal accounts that are from the primary SRP and are from the HD keyring', async () => {
-      getMessenger.mockReturnValue({
-        call: jest.fn().mockResolvedValue(['0x123']),
-      });
-
-      const result =
-        await mapInternalAccountsListToPrimarySRPHdKeyringInternalAccountsList(
-          internalAccountsList,
-          { getMessenger, getUserStorageControllerInstance },
-        );
-
-      expect(result).toStrictEqual([internalAccountsList[0]]);
     });
   });
 });