+++ packages/profile-sync-controller/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add account syncing support for multiple SRPs ([#5753](https://github.com/MetaMask/core/pull/5753))
+
 ## [15.0.0]
 
 ### Changed